Add broken light repair recipes
