+++ src/main/java/dev/theatricalmod/theatrical/data/TheatricalLootTables.java
@@ -1,0 +1,32 @@
+package dev.theatricalmod.theatrical.data;
+
+import com.google.common.collect.ImmutableList;
+import com.mojang.datafixers.util.Pair;
+import net.minecraft.data.DataGenerator;
+import net.minecraft.data.LootTableProvider;
+import net.minecraft.loot.LootParameterSet;
+import net.minecraft.loot.LootParameterSets;
+import net.minecraft.loot.LootTable;
+import net.minecraft.loot.ValidationTracker;
+import net.minecraft.util.ResourceLocation;
+
+import java.util.List;
+import java.util.Map;
+import java.util.function.BiConsumer;
+import java.util.function.Consumer;
+import java.util.function.Supplier;
+
+class TheatricalLootTables extends LootTableProvider {
+    public TheatricalLootTables(DataGenerator dataGeneratorIn) {
+        super(dataGeneratorIn);
+    }
+
+    @Override
+    protected List<Pair<Supplier<Consumer<BiConsumer<ResourceLocation, LootTable.Builder>>>, LootParameterSet>> getTables() {
+        return ImmutableList.of(Pair.of(TheatricalBlockLootTables::new, LootParameterSets.BLOCK));
+    }
+
+    @Override
+    protected void validate(Map<ResourceLocation, LootTable> map, ValidationTracker validationtracker) {
+    }
+}
